codegen: introduce dedicated node kind for magics
Summary
=======

Add the `cnkMagic` node kind, which is meant for referencing magic
procedures, integrate it into the code generators, and adjust `cgirgen`
to use it where possible.

This allows for removing the requirement of producing new `PSym`s for
magics that have no symbol during the MIR phase (i.e. those identified
via `mnkMagic` nodes). For compatibility with the existing
implementation, the `magic` field of symbols is still inspected, but the
direction is to have only `cnkMagic` mean "magic procedure" in the code
generators.

Details
=======

In the code generators, all manual `TSym.magic` lookups, outside of type
contexts, are replaced with calls to either `compat.getMagic` or the new
`compat.getCalleeMagic`. Both procedures consider both `cnkSym` and
`cnkMagic` nodes.

An internal problem that the new node fixes is "improper symbols"
reaching the code generators: a symbol created via `createMagic` is
missing much of what is need for a proper procedure symbol, such as a
type, associated AST, owner, etc. -- with `mnkMagic`, they are no longer
created.