codegen: introduce dedicated node kind for magics (#826)
## Summary

Add the `cnkMagic` node kind, which is meant for referencing magic
procedures,  and integrate it into the code generators. References
to magic procedures that don't use symbols during the MIR phase (i.e.,
`mnkMagic` nodes) don't require the creation of a `PSym` in `cgirgen`
now.

For compatibility with the existing implementation, the `magic` field
of symbols is still inspected during code generation, but the direction
is to have only `cnkMagic` mean "magic procedure" in the code
generators.

## Details

In the code generators, all manual `TSym.magic` lookups, outside of type
contexts, are replaced with calls to either `compat.getMagic` or the new
`compat.getCalleeMagic`. Both procedures consider both `cnkSym` and
`cnkMagic` nodes.

An internal problem that the new node fixes is "improper symbols"
reaching the code generators: a symbol created via `createMagic` is
missing much of what is need for a proper procedure symbol, such as a
type, associated AST, owner, etc. -- with `mnkMagic`, these symbols are
no longer created.

As a side-effect, `render` now uses the enum name for magic-procedure
references using `cnkMagic`; the tests using `--expandArc` are adjusted
to account for this.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>