codegen: lower `swap` via a MIR pass (#828)
## Summary

Implement the lowering of `swap` as a MIR pass instead of directly in
each code generator. This removes the amount of code duplication in the
compiler and fixes a double-evaluation bug for call-expression `swap`
operands when using the JS backend.

MIR passes are now also applied for code run at compile time.

## Details

The MIR passes were previously not applied to code processed by `vmjit`
(compile-time code NimScript execution), which is fixed now.

To be able to efficiently detect and lower the call with a MIR pass,
`mSwap` is transformed into the MIR magic call representation (which
uses `mnkMagic` instead of `mnkCall`). The lowering pass then only has
to search for the `mnkMagic` node and replace it.

Implementation-wise, a `swap` call is lowered into shallow assignments
(`mnkFastAsgn`). For the C target, the code that is later generated
stays the same; for the VM target, the efficiency slightly increases for
operands that aren't locals; and for the JS target, efficiency regresses
for operands of aggregate types. The latter is the case because the
JavaScript code generator currently ignores shallow assignments
(`cnkFastAsgn`) and performs full copies.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>