vmbackend: don't reuse `TCtx`
Don't re-use the `TCtx` type as the orchestrator's context/state type.
Instead, add a dedicated type (`GenCtx`) for this purpose.

This allows for removing the `collectedGlobals` field from `TCtx`, and
it is also a preparation for separating the `vmgen` state from `TCtx`
(the orchestrator needs some place to store the code generator state at,
which it now has with `GenCtx`).

Because it isn't used anywhere else, the `registerProc` procedure is
moved from `vmaux` to `vmbackend` and adjusted for the specific use-
case in the latter.