codegen: optimize single-operand `CgNode`s (#836)
## Summary

Add a dedicated `CgNode` branch for expression-like nodes that always
have a single operand. This reduces the amount of allocations and access
indirections, and it also makes the code more self-describing
(`.operand` now vs. the previous `[0]`).

## Details

For now, only expression-like nodes use the `operand` branch, but some
statement nodes (for example, `mnkRepeat`) could also be considered for
using the `operand` branch in the future. The set previously denoting
all atoms (`cnkWithoutItem`) is renamed to `cnkAtoms` and the usage
sites are adjusted to take the new meaning into account: only traversing
`mnkWithItem` nodes is not enough anymore.

The `cnkCast`, `cnkConv`, and `cnkHiddenConv` nodes also use the
`operand` branch. Them using two subnodes was inherited from `PNode`,
but it's not something that's needed for `CgNode` IR.

`cgirgen` is changed to create `operand`-nodes where required, and all
`[0]` and `[1]` (for `cnkCast|cnkConv|cnkHiddenConv`) access for the
changed nodes is replaced with `.operand`.

### Allocation Stats

The accumulated allocation stats (using `-d:nimAllocStats`) of
`generateIR` when compiling the compiler at
https://github.com/nim-works/nimskull/commit/48a07fbfc214491754761d09b178b71137cc1dd5
are:

|version|allocs|deallocs|
|:------|:-----|:-------|
|this commit|3162007|752474|
|48a07fbfc2|3320434|758663|