remove support for enum inheritance
Summary
=======

Enum types supported an internal base type that could only be specified
in macro-generated enum types (there's no syntax for representing it),
but weren't aren't any tests for it, the features seems to have never
been documented anywhere, and enums that inherit from other enums are
not consistently considered in the compiler. The feature is thus
removed.

Details
=======

- remove support for `nkEnumTy` AST having a non-empty first slot. This
  now results in an `illformed AST` error
- remove support for enum inheritance from `tryReadingTypeField`
- remove the `rsemInheritanceOnlyWorksWithAnEnum` report