vm: pass data for compiler interfacing in `VmArgs`
Summary
=======

Provide the data for interfacing with the compiler in the `VmArgs`
object, which allows for making the callbacks registered by `vmops`
independent of the `TCtx` instance they're first registered with.

Details
=======

Closing over fields of the registered-with-`TCtx` is replaced with
accessing the corresponding `VmArgs` fields.

In addition to making the callbacks context-independent, this also means
that the callbacks in `vmops` aren't (first-class) closures anymore.