vmcompilerserdes: use the correct literal kind
Don't manually create a `nkUIntLit` or `nkIntLit` node, leave this
decision to `newIntLitNode`. As a consequence, the kinds of int literal
nodes coming out of a VM execution are now consistent with what
`semfold` would produce.

This removes another source-of-truth regarding int literal kinds, and is
a preparation for deciding the underlying type of an enum in `sem`.