vm: pass data for compiler interfacing in `VmArgs` (#838)
## Summary

Provide the data for interfacing with the compiler in the `VmArgs`
object, which allows for making the callbacks registered by `vmops`
independent of the `TCtx` instance they're first registered with.

## Details

Closing over fields of the registered-with-`TCtx` is replaced with
accessing the corresponding `VmArgs` fields.

While the `ConfigRef` instance stored with `TCtx` is always the same
as the one stored in `TCtx`'s `graph`, a `ModuleGraph` is not always
present (for example, when the VM runs in standalone mode), meaning
that the `ConfigRef` instance has to also be passed in `VmArgs`.

In addition to making the callbacks context-independent, this change 
also means that the callbacks in `vmops` aren't (first-class) closures
anymore. To prevent them from accidentally capturing something again,
all manual callback procedures in `vmops` are assigned the `.nimcall`
calling convention.