codegen: remove decision making regarding enums
The code generators now use the sem-provided storage type for enums
instead of computing it themselves. Aside from removing yet-another
source-of-truth, this also fixes the internal inconsistency between
the VM and C targets regarding the 32- and 64-bit enum value signed-
ness.

For `vmgen`, `tyEnum` is added to the set of irrelvant types and the
set is used in more places. An exception is made in `genToStr`, where
the `tyEnum` type, not the storage type, is required.