semtypes: select the enum's storage type on definition
The underlying storage type for the enum is now selected when producing
the `tyEnum` type. For back- and forward-compatibility reasons, the
storage type is stored in the `tyEnum`s second slot (allowing `tyEnum`
to be used with `skipTypes`) instead of in the first slot, which is
reserved for a base type.

The selected storage type matches the type that `ccgtypes` previously
selected.