vmops: decouple registration from discovery (#841)
## Summary

Separate *discovering* overrides from *registering* them by turning the
`register` procedures into iterators that don't require a `TCtx`. This
makes the interface more flexibile, and is a preparation for moving the
the `callbackKeys` list out of `TCtx`.

## Details

As part of the changes, the used terminology is improved: a *callback*
is what the VM invokes in the end, while an *override* is a pattern +
callback to redirect to. The places where the `register` procedures
were previously used are updated to iterate the contents of the new
iterators and register the callbacks/overrides as needed.

The `registerX` procedures are turned into iterators, and the calls to
`registerCallback` in `vmops` are replaced with calls to the `override`
template, which expands to a `yield` statement. Iterators are used
instead because they're slightly more flexible, but, at least with the
current usage, constant `seq`s would have worked as well.

Although not yet clear from data placement, "overrides" are not
something that the VM implements -- rather, it's a feature of the
JIT and linker. Therefore, the `registerCallback` is moved from the
`vmdef` into the `vmjit` module, and updated to also verify that the
callbacks and patterns list stay synchronized. The
`registerAdditionalOps`, which has the intended use of registering
the additional ops for compile-time code execution is moved to
`compilerbridge`.

As the result, `vmops` is now only concerned with providing overrides,
but not with how/if they're registered.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>