`std/json`: remove macro type API usage (#825)
## Summary

Remove usage of `macros` type API usage in `std/json`, specifically
`getTypeImpl` and `getTypeInst`. This is part of a broader effort to
drop this API from the standard library. The current API enforces a
number of poor design decisions in compiler internals. The API itself
is questionable and much of this work should happen via generics.

## Details

As part of removing the usage of these APIs, there was a slight
regression in the precision of errors when a tuple without named fields
is used. The error message is equivalent, but the error location will
be within the `json` module itself.

The change itself uses the `fieldPairs` iterator which relies upon
compiler support in `semfields`. Some minor adjustments were made here
in particular the ability to handle `nkError` nodes.

The `json` module relies upon the ability to forward declare generic
procedures, this was not possible with the earlier bootstrap compilers,
and existed behind a conditional compilation flag
(`nimFixedForwardGeneric`). Seeing as this is no longer the case the
flag has been removed as part of the module rework.

Finally, the test `tjsonmacro_reject` existed for the previous approach,
but the test are combined into `tjsonmacro` and the previous test is
removed.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>