generics: fix coercions to distinct phantom types
Summary
=======

Fix phantom type information being lost in conversions-with-inference
(i.e, `Typ(x)` where `Typ` is a generic type) to distinct types.

Details
=======

Matching a `B[U]` against a `A[T] = distinct T`, which is what
eventually happens for conversions like `A(B)`, results in the actual
type `B[U]` being matched against the formal type `distinct T`. Since
the invocation information (i.e., `tyGenericInst`) of the actual
type is skipped when the formal type is a `tyDistinct`, phantom type
information present on the instance was lost.

Skipping the invocation information is now removed, and phantom types
are thus passed on properly.