semtypes: don't sem potentially generic range/index expressions
Range expressions and expressions in the array-index type position are
now first sem'ed as generic expressions, and only if they don't depend
on type variables are they fully semantically analyzed.

This is going to require some changes to the type variables replacing
logic in `semtypinst`. As a preparation for this, the AST coming out of
the generic pre-pass is pre-processed to have symbols bound for all type
variables.