internal: faster `ConfigRef` access (#855)
## Summary

Replace the accessor procedures for `ConfigRef` with templates,
resulting in a small improvement to the compiler's performance.

With a `-d:release` compiler, the time it takes to build the compiler
(excluding the C compilation step) is reduced by ~6% (14.4 compared to
15.3 seconds on the measured-with machine).

## Details

The active configuration accessor procedures for `ConfigRef` didn't use
`lent` (for the access wrappers) and `sink` (for assignment wrapper)
annotations, meaning that each read access introduced an unnecessary
full copy and each assignment introduced a potentially unnecessary full
copy.

Turning the procedures into templates showed to result in slightly
faster code compared to using `lent`, `sink`, and `.inline`, hence them
being turned into templates. Using the accessors is now as fast as
manually accessing the `active` configuration state, and no unnecessary
copies are introduced anymore.